services/app/requirements.txt
flask==2.3.2 
gunicorn==20.1.0
python-dotenv==1.0.0
Flask-WTF==1.1.1
sqlalchemy==2.0.16
email_validator==2.0.0